feat: 데이터 페칭 커스텀훅 수정
