+++ src/hooks/useFetch.js
@@ -1,0 +1,28 @@
+import { useEffect, useState } from "react";
+import mockService from "../services/mockService";
+
+const useFetch = (endpoint) => {
+  const [data, setData] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  const fetchData = async () => {
+    try {
+      const service = mockService;
+      const response = await service.get(endpoint);
+      setData(response.data);
+    } catch (error) {
+      setError(error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchData();
+  }, [endpoint]);
+
+  return { data, loading, error, fetchData };
+};
+
+export default useFetch;
